Move BaseDirective from directive.base to directive
